API Documentation

Below is the API documentation for CodeInfoTools.jl

CodeInfoTools.code_info — Function

code_info(f, tt; generate = true, debuginfo = :default)

Return lowered code for function f with tuple type tt. Equivalent to InteractiveUtils.@code_lowered – but a function call and requires a tuple type tt as input.

Missing docstring.

Missing docstring for Builder. Check Documenter's build log for details.

Missing docstring.

Missing docstring for prepare_builder!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for getindex. Check Documenter's build log for details.

Missing docstring.

Missing docstring for iterate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for circshift!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for bump!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for slump!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for push!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pushfirst!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for insert!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for replace!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for deleteat!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for pushslot!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for finish. Check Documenter's build log for details.