API Documentation

Below is the API documentation for CodeInfoTools.jl

CodeInfoTools.Builder — Type

struct Builder
    ref::CodeInfo
    code::Vector{Any}
    nargs::Int32
    codelocs::Vector{Int32}
    newslots::Dict{Int,Symbol}
    slotnames::Vector{Symbol}
    changemap::Vector{Int}
    slotmap::Vector{Int}
end

Builder(ci::CodeInfo, nargs::Int; prepare=true)

An immutable wrapper around CodeInfo which allows a user to insert statements, change SSA values, insert Core.SlotNumber instances, etc – without effecting the wrapped CodeInfo instance.

Call finish(b::Builder) to produce a modified instance of CodeInfo.

Base.getindex — Function

getindex(b::Builder, i::int)

Return the expression/node at index i from the Vector{Any} of lowered code statements.

Base.iterate — Function

iterate(b::Builder, (ks, i) = (1 : length(b.code), 1))

Iterate over the builder – generating a tuple (k, stmt) at each iteration step, where k is an index and stmt is a node or Expr instance.

CodeInfoTools.code_info — Function

code_info(f, tt; generate = true, debuginfo = :default)

Return lowered code for function f with tuple type tt.

Base.circshift! — Function

circshift!(b::Builder, deg::Int; ch = r -> true, slots::Bool = false)

Shift either the SSA values (slots = false) or the Core.SlotNumber instances (slots = true) selected by ch by deg.

CodeInfoTools.bump! — Function

bump!(b::Builder, v::Int; slots = false)

Increase either the SSA values (slots = false) or the Core.SlotNumber instances for which id >= v by 1.

CodeInfoTools.pushslot! — Function

pushslot!(b::Builder, slot::Symbol)

Insert a new slot into the IR with name slot. Increments all SSA value instances to preserve the correct ordering.

Base.pushfirst! — Function

pushfirst!(b::Builder, stmt)

Push a statement to the head of b.code. This call first shifts all SSA values up by 1 to preserve ordering.

Base.insert! — Function

insert!(b::Builder, v::Int, stmt)

Insert an Expr or node stmt at location v in b.code. Shifts all SSA values with id >= v to preserve order.

Base.replace! — Function

replace!(b::Builder, v::Int, stmt)

Replace the Expr or node at location v with stmt.

CodeInfoTools.prepare_builder! — Function

prepare_builder!(b::Builder)

Iterate over the reference CodeInfo instance in b.ref – pushing Expr instances and nodes onto the builder b. This function is called during Builder construction so that the user is presented with a copy of the CodeInfo in b.ref.

CodeInfoTools.finish — Function

finish(b::Builder)

Produce a new CodeInfo instance from a Builder instance b.