API Documentation

Below is the API documentation for CodeInfoTools.jl

CodeInfoTools.code_info — Function

code_info(f::Function, tt::Type{T}; generated = true, debuginfo = :default) where T <: Tuple
code_info(f::Function, t::Type...; generated = true, debuginfo = :default)

Return lowered code for function f with tuple type tt. Equivalent to InteractiveUtils.@code_lowered – but a function call and requires a tuple type tt as input.

CodeInfoTools.Variable — Type

const Variable = Core.SSAValue
var(id::Int) = Variable(id)

Alias for Core.SSAValue – represents a primitive register in lowered code. See the section of Julia's documentation on lowered forms for more information.

CodeInfoTools.Statement — Type

struct Statement{T}
    node::T
    type::Any
end

A wrapper around Core nodes with an optional type field to allow for user-based local propagation and other forms of analysis. Usage of Builder or Canvas will automatically wrap or unwrap nodes when inserting or calling finish – so the user should never see Statement instances directly unless they are working on type propagation.

For more information on Core nodes, please see Julia's documentation on lowered forms.

CodeInfoTools.Canvas — Type

struct Canvas
    defs::Vector{Tuple{Int, Int}}
    code::Vector{Any}
    codelocs::Vector{Int32}
end
Canvas() = Canvas(Tuple{Int, Int}[], [], Int32[])

A Vector-like abstraction for Core code nodes.

Properties to keep in mind:

Insertion anywhere is slow.

Pushing to beginning is slow.

Pushing to end is fast.

Deletion is fast.

Accessing elements is fast.

Setting elements is fast.

Thus, if you build up a Canvas instance incrementally, everything should be fast.

CodeInfoTools.Builder — Type

Builder(ir)

A wrapper around a Canvas instance. Call finish when done to produce a new CodeInfo instance.

Base.iterate — Function

iterate(b::Builder, (ks, i) = (pipestate(p.from), 1))

Iterate over the original CodeInfo and add statements to a target Canvas held by b::Builder. iterate builds the Canvas in place – it also resolves local GlobalRef instances to their global values and inlines them at the head of Expr(:call, ...) instances. iterate is the key to expressing idioms like:

for (v, st) in b
    b[v] = swap(st)
end

At each step of the iteration, a new node is copied from the original CodeInfo to the target Canvas – and the user is allowed to setindex!, push!, or otherwise change the target Canvas before the next iteration. The naming of Core.SSAValues is taken care of to allow this.

CodeInfoTools.finish — Function

finish(b::Builder)

Create a new CodeInfo instance from a Builder. Renumbers the wrapped Canvas in-place – then copies information from the original CodeInfo instance and inserts modifications from the wrapped Canvas