API Documentation

Below is the API documentation for CodeInfoTools.jl

CodeInfoTools.code_info — Function

code_info(f, tt; generate = true, debuginfo = :default)

Return lowered code for function f with tuple type tt. Equivalent to InteractiveUtils.@code_lowered – but a function call and requires a tuple type tt as input.

CodeInfoTools.Pipe — Type

Pipe(ir)

A wrapper around a Canvas object. Call finish when done to produce a new CodeInfo instance.

Missing docstring.

Missing docstring for renumber. Check Documenter's build log for details.

Missing docstring.

Missing docstring for renumber!. Check Documenter's build log for details.

CodeInfoTools.finish — Function

finish(p::Pipe)

Create a new CodeInfo instance from a Pipe. Renumbers the wrapped Canvas in-place – then copies information from the original CodeInfo instance and inserts modifications from the wrapped Canvas.