API Documentation

Below is the API documentation for CodeInfoTools.jl

CodeInfoTools.code_info — Function

code_info(f, tt; generate = true, debuginfo = :default)

Return lowered code for function f with tuple type tt. Equivalent to InteractiveUtils.@code_lowered – but a function call and requires a tuple type tt as input.

CodeInfoTools.Builder — Type

struct Builder
    ref::CodeInfo
    code::Vector{Any}
    nargs::Int32
    codelocs::Vector{Int32}
    newslots::Dict{Int,Symbol}
    slotnames::Vector{Symbol}
    slotmap::Vector{Int}
end

Builder(ci::CodeInfo, nargs::Int; prepare=true)

An immutable wrapper around CodeInfo which allows a user to insert statements, change SSA values, insert Core.SlotNumber instances, etc – without effecting the wrapped CodeInfo instance.

Call finish(b::Builder) to produce a modified instance of CodeInfo.

CodeInfoTools.prepare_builder! — Function

prepare_builder!(b::Builder)

Iterate over the reference CodeInfo instance in b.ref – pushing Expr instances and nodes onto the builder b. This function is called during Builder construction so that the user is presented with a copy of the CodeInfo in b.ref.

Base.getindex — Function

getindex(b::Builder, i::Int)
getindex(b::Builder, i::Core.SSAValue)

Return the expression/node at index i from the Vector of lowered code statements.

Base.iterate — Function

iterate(b::Builder, (ks, i) = (1 : length(b.code), 1))

Iterate over the builder – generating a tuple (Core.SSAValue(k), stmt) at each iteration step, where k is an index and stmt is a node or Expr instance.

Base.circshift! — Function

circshift!(b::Builder, deg::Int; ch = r -> true, slots::Bool = false)

Shift either SSA values (slots = false) or the Core.SlotNumber instances (slots = true) by deg. The Boolean function ch determines which subset of values are shifted and can be customized by the user.

CodeInfoTools.bump! — Function

bump!(b::Builder, v::Int; slots = false)

Subsets all instances of Core.SSAValue or Core.SlotNumber greater than v and shifts them up by 1. Convenience form of circshift!.

CodeInfoTools.slump! — Function

slump!(b::Builder, v::Int; slots = false)

Subsets all instances of Core.SSAValue or Core.SlotNumber greater than v and shifts them down by 1. Convenience form of circshift!.

Base.push! — Function

push!(b::Builder, stmt)

Push a statement to the end of b.code.

Base.pushfirst! — Function

pushfirst!(b::Builder, stmt)

Push a statement to the head of b.code. This call first shifts all SSA values up by 1 to preserve ordering.

Base.insert! — Function

insert!(b::Builder, v::Int, stmt)
insert!(b::Builder, v::Core.SSAValue, stmt)

Insert an Expr or node stmt at location v in b.code. Shifts all SSA values with id >= v to preserve order.

Base.replace! — Function

replace!(b::Builder, v::Int, stmt)
replace!(b::Builder, v::Core.SSAValue, stmt)

Replace the Expr or node at location v with stmt.

Base.deleteat! — Function

deleteat!(b::Builder, v::Int)
deleteat!(b::Builder, v::Core.SSAValue)

Delete the expression or node at location v. If v indexes a Core.NewvarNode (which indicates a slot), the slotname is also removed from b.slotnames. All SSA values and slots are shifted down accordingly.

CodeInfoTools.pushslot! — Function

pushslot!(b::Builder, slot::Symbol)

Insert a new slot into the IR with name slot. Increments all SSA value instances to preserve the correct ordering.

CodeInfoTools.finish — Function

finish(b::Builder)

Produce a new CodeInfo instance from a Builder instance b.